+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/work/WorkChore.kt
@@ -1,12 +1,7 @@
 package com.altmann.choresmanager.models.chores.work
 
-import androidx.compose.ui.graphics.Color
-import com.altmann.choresmanager.models.Priority
 import com.altmann.choresmanager.models.chores.Chore
-import kotlinx.datetime.DateTimePeriod
-import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
-import kotlinx.datetime.LocalTime
 
 class WorkChore(
     chore : Chore,