+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/calendar/CalendarEvent.kt
@@ -7,7 +7,7 @@ sealed interface CalendarEvent {
     data class SelectDate(val date: LocalDate) : CalendarEvent
     object DismissExpanded : CalendarEvent
     data class AddChore(val chore: Chore) : CalendarEvent
-    data class MarkFinished(val choreId : Int, val date : LocalDate) : CalendarEvent
+    data class MarkFinished(val choreId : String, val date : LocalDate) : CalendarEvent
     object LoadChores : CalendarEvent
     object PrevMonth : CalendarEvent
     object NextMonth : CalendarEvent