+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/viewmodels/SharedChoreViewModel.kt
@@ -2,8 +2,9 @@ package com.altmann.choresmanager.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.altmann.User
+import com.altmann.choresmanager.models.User
 import com.altmann.choresmanager.models.chores.Chore
+import com.altmann.choresmanager.utils.AchievementHelper
 import com.altmann.choresmanager.utils.CalendarHelper
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -27,6 +28,11 @@ class SharedChoreViewModel : ViewModel() {
     private val _chores = MutableStateFlow<List<Chore>>(emptyList())
     val chores = _chores.asStateFlow()
 
+    private val _user = MutableStateFlow(
+        User(userId = 1, name = "Michel", birthday = LocalDate(2004, 9, 5), profileImage = null)
+    )
+    val user = _user.asStateFlow()
+
     private val _enabledChores = MutableStateFlow<List<Chore>>(emptyList())
     val enabledChores = _enabledChores.asStateFlow()
 
@@ -41,7 +47,9 @@ class SharedChoreViewModel : ViewModel() {
     fun addChore(chore: Chore) {
         _chores.value = _chores.value.plus(chore)
         _enabledChores.value = _enabledChores.value.plus(chore)
+        _user.update { it.copy(createdChores = it.createdChores + 1) }
         remapChores()
+        updateAchievements()
     }
 
     fun enableDisableChore(chore: Chore) {
@@ -55,20 +63,53 @@ class SharedChoreViewModel : ViewModel() {
 
     fun onPrev() = _anchor.update { it.minus(DatePeriod(months = 1)) }
 
-    fun markChoreFinished(choreId: Int, date : LocalDate) {
-        _chores.value.find { it.choreId == choreId }?.let { chore ->
+    fun markChoreFinished(choreId: String, date: LocalDate) {
+        _chores.value.find { it.choreId == choreId}?.let { chore ->
             if (date == chore.endDate) {
                 chore.finishChore()
                 _chores.value = _chores.value.map { if (it.choreId == choreId) chore else it }
                 _enabledChores.value = _enabledChores.value.filter { it.choreId != choreId }
+                // Update user completed chores
+                addCompletedChoreToUser(chore)
             } else {
                 chore.choreException = chore.choreException.plus(date)
                 _chores.value = _chores.value.map { if (it.choreId == choreId) chore else it }
-                _enabledChores.value = _enabledChores.value.map { if (it.choreId == choreId) chore else it }
+                _enabledChores.value =
+                    _enabledChores.value.map { if (it.choreId == choreId) chore else it }
             }
         }
         // Keep enabled list; finished chores will be excluded by mapping predicate
         remapChores()
+
+        // Update achievements
+        updateAchievements()
+    }
+
+    fun addCompletedChoreToUser(chore: Chore) {
+        _user.value = _user.value.copy(
+            completedChores = _user.value.completedChores.plus(chore)
+        )
+        _user.update { it.gainExp(10) }// Gain 10 XP per completed chore
+    }
+
+    fun levelUpUser() {
+        _user.value.levelUp()
+    }
+
+    private fun updateAchievements() {
+        AchievementHelper(user.value.achievements, completedChores = user.value.completedChores, createdChores = user.value.createdChores)
+            .checkForNewAchievements()
+            .let { (newAchievements, xp) ->
+                if (newAchievements.isNotEmpty()) {
+                    _user.update { currentUser ->
+                        currentUser.copy(
+                            achievements = currentUser.achievements.plus(newAchievements)
+                        )
+                    }
+                }
+                _user.update { it.gainExp(xp) }
+                print("Gained $xp XP from achievements: $newAchievements")
+            }
     }
 
     private fun generateChoreMap(